CMS-2111 Exception when trying to update a page template with a new XSL

+++ modules/cms-core/src/main/java/com/enonic/cms/core/structure/page/template/PageTemplateServiceImpl.java
@@ -622,7 +622,10 @@ public class PageTemplateServiceImpl
             if ( updatedPTPs == 0 )
             {
                 Integer[] pageKeys = getPageKeysByPageTemplateKey( pageTemplateKey );
-                pageWindowDao.deleteByPageKeys( pageKeys );
+                if ( pageKeys.length != 0 )
+                {
+                    pageWindowDao.deleteByPageKeys( pageKeys );
+                }
 
                 pageTemplate.clearPageTemplateRegions();
             }
